"check_style_score -> 5 of 5 passed. style score: 0/1 "

+++ CSPP2-assignments/m2/Assignment-3/Solution.java
@@ -6,15 +6,15 @@ public class Solution {
 	*/
 	public static void main(String[] args) {
 		Scanner s=new Scanner(System.in);      
-        int base = s.nextInt();
+        long base = s.nextLong();
         int exponent = s.nextInt();
-        int result=power(base,exponent);
+        long result=power(base,exponent);
         System.out.println(result);
 	}
 	/*
 	Need to write the power function and print the output.
 	*/
-	public static int power(int base, int exponent) {
+	public static long power(long base, int exponent) {
 		if (exponent == 0 ){
 			return(1);
 		} else {
